+++ MigrationTools.sln
@@ -16,8 +16,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Chocolatey", "Chocolatey", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VstsSyncMigrator.Core.Tests", "src\VstsSyncMigrator.Core.Tests\VstsSyncMigrator.Core.Tests.csproj", "{AB3922F0-E801-474A-A498-892C3EA99799}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VstsSyncMigrator.Console.Tests", "src\VstsSyncMigrator.Console.Tests\VstsSyncMigrator.Console.Tests.csproj", "{985259EA-C475-4918-9342-50AB744F52AF}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Extension", "Extension", "{243265E4-0E5A-4E95-BE7C-B92B7666FF79}"
 	ProjectSection(SolutionItems) = preProject
 		src\VstsSyncMigrator.Extension\LICENSE.txt = src\VstsSyncMigrator.Extension\LICENSE.txt
@@ -113,10 +111,6 @@ Global
 		{AB3922F0-E801-474A-A498-892C3EA99799}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AB3922F0-E801-474A-A498-892C3EA99799}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AB3922F0-E801-474A-A498-892C3EA99799}.Release|Any CPU.Build.0 = Release|Any CPU
-		{985259EA-C475-4918-9342-50AB744F52AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{985259EA-C475-4918-9342-50AB744F52AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{985259EA-C475-4918-9342-50AB744F52AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{985259EA-C475-4918-9342-50AB744F52AF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A314A07F-1D5E-44B8-9973-3AB2C41F17D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A314A07F-1D5E-44B8-9973-3AB2C41F17D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A314A07F-1D5E-44B8-9973-3AB2C41F17D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -154,7 +148,6 @@ Global
 		{68BAF6BA-2BC0-48D0-B892-1D9F93D7003A} = {E5294FF3-167A-49A5-A453-19332DA32B01}
 		{89434800-3004-4621-8677-C14C1D1AA4AF} = {32E2956A-9056-4D9D-8018-14DFD0E6CA41}
 		{AB3922F0-E801-474A-A498-892C3EA99799} = {E5294FF3-167A-49A5-A453-19332DA32B01}
-		{985259EA-C475-4918-9342-50AB744F52AF} = {D0EE5917-21B2-49FA-9506-69857EE82DCD}
 		{243265E4-0E5A-4E95-BE7C-B92B7666FF79} = {32E2956A-9056-4D9D-8018-14DFD0E6CA41}
 		{2BC0808B-6F4D-4635-82FE-1118181DD64F} = {A05A7847-4C75-4E92-B156-F529483688CC}
 		{E9399C6D-A5E0-43ED-8412-6C4A6BC86EC8} = {2BC0808B-6F4D-4635-82FE-1118181DD64F}
