+++ src/MigrationTools.Tests/Core/Configuration/EngineConfigurationBuilderTests.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using MigrationTools.Core.Configuration;
+using System.IO;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
@@ -48,6 +49,29 @@ namespace MigrationTools.Core.Configuration.Tests
         {
             var ecb = CreateEngine();
             ecb.BuildWorkItemMigration();
+        }
+
+        [TestMethod]
+        public void TestSeraliseToJson()
+        {
+
+            HelperCreateDefaultConfigFile();
+
+        }
+
+        [TestMethod]
+        public void TestDeseraliseFromJson()
+        {
+            HelperCreateDefaultConfigFile();
+             EngineConfiguration ec;
+            StreamReader sr = new StreamReader("configuration.json");
+            string configurationjson = sr.ReadToEnd();
+            sr.Close();
+            ec = JsonConvert.DeserializeObject<EngineConfiguration>(configurationjson,
+                new FieldMapConfigJsonConverter(),
+                new ProcessorConfigJsonConverter());
+            Assert.AreEqual(10, ec.FieldMaps.Count);
+            Assert.AreEqual(13, ec.Processors.Count);
         }
     }
 }
