--- src/VstsSyncMigrator.Console.Tests/ApplicationInsights.config
@@ -1,32 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ApplicationInsights xmlns="http://schemas.microsoft.com/ApplicationInsights/2013/Settings">
-	<TelemetryModules>
-		<Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.PerformanceCollectorModule, Microsoft.AI.PerfCounterCollector">
-			<!--
-      Use the following syntax here to collect additional performance counters:
-      
-      <Counters>
-        <Add PerformanceCounter="\Process(??APP_WIN32_PROC??)\Handle Count" ReportAs="Process handle count" />
-        ...
-      </Counters>
-      
-      PerformanceCounter must be either \CategoryName(InstanceName)\CounterName or \CategoryName\CounterName
-      
-      NOTE: performance counters configuration will be lost upon NuGet upgrade.
-      
-      The following placeholders are supported as InstanceName:
-        ??APP_WIN32_PROC?? - instance name of the application process  for Win32 counters.
-        ??APP_W3SVC_PROC?? - instance name of the application IIS worker process for IIS/ASP.NET counters.
-        ??APP_CLR_PROC?? - instance name of the application CLR process for .NET counters.
-      -->
-		</Add>
-		<Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse.QuickPulseTelemetryModule, Microsoft.AI.PerfCounterCollector"/>
-	</TelemetryModules>
-	<TelemetrySinks>
-		<Add Name="default">
-			<TelemetryProcessors>
-				<Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse.QuickPulseTelemetryProcessor, Microsoft.AI.PerfCounterCollector"/>
-			</TelemetryProcessors>
-		</Add>
-	</TelemetrySinks>
-</ApplicationInsights>
